requirements.txt
autopep8
pycodestyle
PyPDF2
tomli
typing_extensions
fastapi
uvicorn
google-genai
python-dotenv